Don't scan for a response after issuing a 'join'.
Scanning for the response to 'join' will usually timeout because the
response arrives after the DHCP lease is obtained.

- UART timeout: 500ms
- typical delay to obtain an IP via DHCP: ~3000ms

When the response arrives it is copied into the UART RX buffer. If the
program carries on without clearing this buffer and sends an HTTP request
over the TCP socket, it will behave as if the response was instantaneous,
because there will be available data in the RX buffer.

It will then start to parse this data, and if the actual HTTP response
takes longer than the UART timeout to arrive (which is not the most
frequent case but still happens quite often) the program will simply
consider the request to have failed and discard the valid response.