Connect to a public Pusher channel.
We cannot use client events the regular way because we can only have one
socket open at a time and this prevents us from performing the 3-tier
handshake usually required (the Pusher server + the reaDIYmate server +
the client device).

To trigger client events on a private channel, we need to hardcode the
private key for the Pusher API on the device. However this is not viable
with many devices, for security reasons.

Therefore we trigger server events on a public channel with a random name,
from the iPhone. The client device simply retrieves the channel name and
receives the server events.